requirements.txt
flask==2.3.2
couchbase>=4.3.0,<4.4.0
faker==18.9.0
boto3>=1.35.0
python-dotenv>=1.0.0
streamlit>=1.28.0
